make HelperDetail components
