+++ app/src/main/java/com/example/helpersapp/ui/screens/HelperDetailsScreen.kt
@@ -1,12 +1,8 @@
 package com.example.helpersapp.ui.screens
 
-import android.content.ContentValues
 import android.util.Log
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
@@ -14,7 +10,6 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.ArrowBack
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -28,22 +23,17 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import coil.compose.rememberAsyncImagePainter
 import com.example.helpersapp.model.HelperInfo
 import com.example.helpersapp.ui.components.ShowBottomImage
-import com.example.helpersapp.ui.components.createUsername
+import com.example.helpersapp.ui.components.HelperDetail
 import com.example.helpersapp.viewModel.HelperViewModel
 import com.example.helpersapp.viewModel.LoginViewModel
-import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.ktx.storage
-
-
 
 @Composable
 fun HelperDetailsScreen(
@@ -165,24 +155,5 @@ fun HelperDetailsScreen(
                 }
             }
         }
-    }
-}
-
-@Composable
-fun HelperDetail(helperInfo: HelperInfo) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(bottom = 50.dp, start = 16.dp, end = 16.dp, top = 16.dp),
-        verticalArrangement = Arrangement.spacedBy(4.dp)
-    ) {
-        Spacer(modifier = Modifier.height(15.dp))
-        Text(text = "About:\n${helperInfo.about}")
-        Spacer(modifier = Modifier.height(15.dp))
-        Text(text = "Category:\n${helperInfo.category}")
-        Spacer(modifier = Modifier.height(15.dp))
-        Text(text = "Details:\n${helperInfo.details}")
-        Spacer(modifier = Modifier.height(15.dp))
-        Text(text = "Experience:\n${helperInfo.experience}")
     }
 }
